requirements.txt
# Core HTTP & scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF processing
pypdf>=3.0.0

# Google Gemini (AI analysis of Diário Oficial)
google-generativeai>=0.7.0

# Geospatial & KMZ generation
geopandas>=0.14.0
shapely>=2.0.0
simplekml>=1.3.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# OpenAI-compatible clients (multi-model fallback)
openai>=1.0.0

# Supabase (vertices database)
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0
